Move around methods to fix cyclic dependency
